\name{cmdata}
\alias{cmdata}
\title{
	Protein structures for user examples.
}
\description{
	Protein structures used to produce figures found in Schweppe et al.. Including structures for Oxa23 (4G4Y), Ab57_2983 (Phyre2), PPIB (1CYN), PPIB-decoy, and TbpA (3P9D).
}
\details{
\tabular{ll}{
	Package: \tab XLMap\cr
	Type: \tab Dataset\cr
	Version: \tab 0.1\cr
	Date: \tab 2015-06-15\cr
	License: \tab GPL-2\cr
}
	".pdb" files for creating contact map files.
	Currently five structures exist: Oxa23 (4G4Y), Ab57_2983 (Phyre2), PPIB (1CYN), PPIB-decoy, and TbpA (3P9D).
	
}
\author{
Devin K Schweppe, PhD
Maintainer: Devin K Schweppe <dkschwep@uw.edu>
}
\references{
Schweppe et al.
}
\keyword{ structure data}

\alias{p.oxa23}
\alias{p.ab57.2983}
\alias{p.ppib}
\alias{p.ppib.decoy}
\alias{p.tbpa}