\name{read.pdbx.atoms}
\alias{read.pdbx.atoms}
\title{
Read the ATOMS subsection from PDBx/mmCIF files.
}
\description{
An extension of the 'read.pdb' function from Rpdb, this function can import atom information from PDBx/mmCIF files. 
}
\usage{
	read.pdbx.atoms(file)
}

\arguments{
  \item{file}{
    A PDBx/mmCIF file, file extension '.cif'.
}
}
\details{
	Required input is A PDBx/mmCIF file, file extension '.cif'.
}
\references{
Schweppe et al.
}
\author{
Devin K Schweppe, Bruce Lab, University of Washington
}