\name{xlmapdata}
\alias{xlmapdata}
\title{
	Cross-linking data for user examples.
}
\description{
	Cross-linking data used to produce figures found in Schweppe et al.. Including data for Oxa23, Ab57_2983, PPIB, and TbpA (data for TbpA comes from Leitner et al. 2012).
}
\details{
\tabular{ll}{
	Package: \tab XLMap\cr
	Type: \tab Dataset\cr
	Version: \tab 0.1\cr
	Date: \tab 2015-06-15\cr
	License: \tab GPL-2\cr
}
	".pdb" files for creating contact map files.
	Currently data for four proteins are included: Oxa23, Ab57_2983, PPIB, and TbpA.
	
}
\author{
Devin K Schweppe, PhD
Maintainer: Devin K Schweppe <dkschwep@uw.edu>
}
\references{
Schweppe et al.
}
\keyword{ crosslink data}

\alias{oxa23}
\alias{ab57.2983}
\alias{ppib}
\alias{tbpa}